@@ -79,6 +79,6 @@ module.exports = {
         "import/no-extraneous-dependencies": "off",
         "@typescript-eslint/no-explicit-any": "off"
       }
-    }
+    },
   ]
 };
